+++ src/meshProvider.ts
@@ -125,6 +125,7 @@ export class MeshViewProvider
       pointColor: config.get("pointColor", "#cc0000"),
       wireframeColor: config.get("wireframeColor", "#0000ff"),
       fogDensity: config.get("fogDensity", 0.01),
+      lightIntensity: config.get("lightIntensity", 1.0),
       cameraControls: config.get("cameraControls", "trackball"),
     };
     return `<meta id="vscode-3dviewer-data" data-settings="${JSON.stringify(